refactor: Inline trips.ts into index.ts and move getDatedServiceJourney
